Provide utilities for common functions
