+++ sansa-ml-common/src/main/scala/net/sansa_stack/ml/common/outliers/vandalismdetection/feature/Utils.scala
@@ -1,9 +1,8 @@
 package net.sansa_stack.ml.common.outliers.vandalismdetection.feature
 
-import java.util.regex.{ Pattern
-import org port org.apache.spark.sql._
+import java.util.regex.{ Matcher, Pattern }
 
-object FacilitiesClass extends Serializable {
+object Utils extends Serializable {
 
   def cleaner(str: String): String = {
     val cleaned_value1 = str.replace("{", "").trim()
@@ -16,100 +15,13 @@ object FacilitiesClass extends Serializable {
     namesList
   }
 
-  // ok --- Used for DF Triples
-  def RDD_TO_DFR_RDFXML(rdd: RDD[String], sqlContext: org.apache.spark.sql.SQLContext): DataFrame = {
-    // Create an Encoded Schema in a String Format:
-    val schemaString = "Subject Predicate Object"
-    // Generate schema:
-    val schema = StructType(schemaString.split(" ").map(fieldName => StructField(fieldName, StringType, true)))
-    // Apply Transformation for Reading Data from Text File
-    val rowRDD = rdd.map(_.split(" ")).map(e => Row(e(0), e(1), e(2)))
-    // Apply RowRDD in Row Data based on Schema:
-    val RDFTRIPLE = sqlContext.createDataFrame(rowRDD, schema)
-    // Store DataFrame Data into Table
-    RDFTRIPLE.registerTempTable("SPO")
-    // Select Query on DataFrame
-    val dfr = sqlContext.sql("SELECT * FROM SPO")
-    dfr.show()
-
-    dfr
-  }
-
-  // ok --- Used for DF Triples
-  def RDD_TO_DFR_TRIX(rdd: RDD[String], sqlContext: org.apache.spark.sql.SQLContext): DataFrame = {
-    // Create an Encoded Schema in a String Format:
-    val schemaString = "Subject Predicate Object"
-    // Generate schema:
-    val schema = StructType(schemaString.split(" ").map(fieldName => StructField(fieldName, StringType, true)))
-    // Apply Transformation for Reading Data from Text File
-    val rowRDD = rdd.map(_.split("><")).map(e => Row(e(0), e(1), e(2)))
-    // Apply RowRDD in Row Data based on Schema:
-    val RDFTRIPLE = sqlContext.createDataFrame(rowRDD, schema)
-    // Store DataFrame Data into Table
-    RDFTRIPLE.registerTempTable("SPO")
-    // Select Query on DataFrame
-    val dfr = sqlContext.sql("SELECT * FROM SPO")
-    dfr.show()
-
-    dfr
-  }
-
-  // ok --- Used for DF Triples
-  def RDD_TO_DFR_JTriple(rdd: RDD[String], sqlContext: org.apache.spark.sql.SQLContext): DataFrame = {
-    // Create an Encoded Schema in a String Format:
-    val schemaString = "Subject Predicate Object"
-    // Generate schema:
-    val schema = StructType(schemaString.split(" ").map(fieldName => StructField(fieldName, StringType, true)))
-    // Apply Transformation for Reading Data from Text File
-    val rowRDD = rdd.map(_.split(",")).map(e => Row(e(0), e(1), e(2)))
-    // Apply RowRDD in Row Data based on Schema:
-    val RDFTRIPLE = sqlContext.createDataFrame(rowRDD, schema)
-    // Store DataFrame Data into Table
-    RDFTRIPLE.registerTempTable("SPO")
-    // Select Query on DataFrame
-    val dfr = sqlContext.sql("SELECT * FROM SPO")
-    dfr.show()
-
-    dfr
-  }
-
   def roundDouble(va: Double): Double =
     Math.round(va * 10000).toDouble / 10000
 
   def stringToInt(str: String): Integer =
     str.toInt
 
-  def arrayString2VectorDouble(str: String): Vector = {
-
-    val vector: Vector = Vectors.zeros(0)
-    val str_recordList: Array[String] = str.split(",")
-    val size: Integer = str_recordList.size
-    var double_recordList: Array[Double] = new Array[Double](size)
-
-    // val x= str_recordList.toVector
-
-    for (record <- str_recordList) {
-
-      if (record.nonEmpty) {
-
-        val tem0: String = record.replace("[", "").trim()
-        val tem1: String = tem0.replace("]", "").trim()
-
-        val tem2 = tem1.toDouble
-        double_recordList +:= tem2
-
-      }
-
-    }
-
-    toVector(double_recordList)
-
-  }
-
-  def toVector(arra: Array[Double]): Vector =
-    Vectors.dense(arra)
-
-  def array2String(arra: Array[Double]): String = {
+  def arrayToString(arra: Array[Double]): String = {
 
     var tem = ""
 
@@ -143,4 +55,18 @@ object FacilitiesClass extends Serializable {
     val result: Double = characterRatio(str, pattern)
     result
   }
+
+  def stringMatch(str: String, patternStr: String): Matcher = {
+    val pattern: Pattern = Pattern.compile(patternStr)
+    val matcher: Matcher = pattern.matcher(str)
+    matcher
+  }
+
+  def stringMatchValue(str: String, patternStr: String): Double = {
+    val matcher = stringMatch(str, patternStr)
+    var count: Double = 0.0
+    while (matcher.find()) { count += 1; count - 1 }
+    count
+  }
+
 }
